feat setup gitApp
